move GenerateUserData to os specific impl
to support user-mode networking in hyperv we need to customize the
cloud-init config. This patch just move GenerateUserData outside of the
default cloudinit impl so it can be customized for hyperv in following
commit.
